requirements_backend.txt
flask>=2.3.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
werkzeug<3.0.0
python-dotenv>=1.0.0
openai>=1.0.0
pillow>=10.0.0
browser-use>=0.5.4
playwright>=1.40.0
requests>=2.31.0 